msp430 chat now checks for 0 word count in read-words
Before it would dump out the contents of memory - twice. A byte-length
word count of zero became a 16-bit zero count; after the first word was
sent back, it was decremented to ffff, and then, after sending 64 Ki
words of memory, it would stop.

Now, when given a byte-length count of zero, it does nothing, which is
as it should be.