Remove LOCKA and streamline flash-word
I was erroneously including the LOCKA bit, thinking I had to write ! to
keep it set... Turns out the writing 1 actually clears it!

So I've removed LOCKA from the equates, so mistakes are harder to make,
and also streamlined the flash-word processing: now it expects the flash
command on the stack to include the a500 high byte (FKEY).