Change the names of some registers, and moved them around on chat frame
So, this is just me being a big dork.

I decided that I wanted to favour readability over terseness, so I
changed the following register names:

  i  ->  ip
  s  ->  sp
  r  ->  rp
  t  ->  top
 cnt ->  count

and I switched count and top, both in the register convention
definitions, and on the chat frame. Why? Who knows. Mostly because I
didn't want top te be r7. I much happier now that it's r8. ;-)

Perhaps I've been working on this code for too long... It's been a long
week!!