I guess config.h was short-lived ;-)
Found a better way to do this that obviates the use of _any_ of the
-DHAS_FOOBAR defines. I also got rid of -DCHUMOO - push and pop are
compiled now, no matter what. If they need to be removed from the
kernel, they easily can be.

I'm also assuming, in this code, that tty support is approximately equal
to termios support, and that termios support, in its turn, suggests
serial port support. Now, by putting tty.o in the archobjs list in
configure.sh, for a particular OS target, you get it all: isatty,
termios, serial support.