Added a "peer" mode for msp chat
Another muforth process can, using a pty, emulate the MSP430 chat
firmware. This can be useful for vetting the protocol, finding silly
bugs, having fun, etc.

I compiled the pty.c code on OSX and was able to get it work, with one
caveat: if the slave side disconnects - say, after loading the
ram-resident chat code, to switch to the flash-resident code -  the
master side of the pty times out, and can not be convinced to start
waiting for more characters (commands) until the slave side sends a
byte.

This is inconvenient and annoying. But it's possible to restart the
conversation and test the whole process of pretending to flash a new
bootloader.