Fixed ReadWord and WriteWord in chat-host
The interface (in interact) shows them taking an address of the word to
read or write, but the implementation didn't take an address - it
assumed a streaming model.

So, let's bridge between the two. If the address that we're reading or
writing is 2 more than the last time, don't SetAddr - just run the
streaming word.

This is untested code! I'm thinking of using the peer implementation to
check it...