Based on MSP430 experience, implement do/loop differently
The "traditional" method - circa FORTH-83, anyway - for implementing
do/loop is to store the limit and index in such a way (with the high bit
toggled) that the end of the loop is signalled by crossing from 7fff to
8000 (or vice versa!) - which signals an arithmetic overflow (ie, sets
the V bit).

Since I didn't have access to the V bit when writing the C
implementations for do/loop/+loop, I instead came up with the idea of
storing the index and limit normally, checking whether the index crosses
_zero_ - ie, if its sign changes - to signal loop end.

In C this is an easy check: xor the limit before incrementing, with the
limit after, and if the xor is < 0, you're done.

When implementing do/loop/+loop on the MSP430, where, in theory, I have
access to the V bit, I started out trying the traditional FORTH-83
method, but then realised: I can't actually branch on the state of the V
bit!! I can only branch on _signed greater than or equal_ (N xor V = 0)
or _signed less than_ (N xor V = 1)

Thus, it made sense to try the zero-crossing method instead, and - ecco
la! - it's nice and simple and efficient.

Given this experience, I changed the S08 kernel code to match. And I'll
probably implement do/loop/+loop this way in the future as well!