Added capability of doing "paranoid" chat - with checksums
Here's the idea:

Rather than totally remaking the chat protocol and starting from
scratch, I decided it would be interesting to try to "retrofit" an
optional checksum step to the protocol.

The way it works is that both the host and target checksum commands and
data as they go by on the wire. The checksum is a 16-bit word; data
words are summed as words, by xoring into the sum. Command bytes, and
byte-sized word counts (for the ReadWords command) are xored in as if
they were words, with the high byte zero.

At the end of a transaction - and that is decided by the host - the host
requests the targets current checksum, and checks it against its own.

Currently SetAddr is checksummed independently of the data that is read
or written afterwards. This way the command can (in theory) be re-tried
if the wrong address is set - but before, eg, writing several words to
the target.

GetSP is treated similarly - the three bytes - command, and SP value -
are checksummed as a unit.

So that's is easier to upgrade existing chat code, paranoid mode is
_off_ by default. To switch it on - and flash a paranoid-capable chat -
do the following:

  ./muforth -d in-ram -f target/MSP430/load-chat.mu4
  chat
  reset
  ^C

You are now running checksum-capable chat code - from RAM.

  ./muforth -d paranoid -f target/MSP430/load-chat.mu4
  chat
  flash-image
  verify

At this point you should either define paranoid in
target/MSP430/build.mu4, or just remember to invoke it when you care
about getting checksums.