Fixed a few checksum bugs
First, when a bad checksum was received, I was always printing ffff
rather than the actual checksum.

Second, when the target ran code, there was the possibility that it
would clobber the Y register - our checksum accumulator. So, chat-peer
clobbers it to emulate this, and now chat-host forces a clear (by
reading the checksum and throwing it away) after doing a target Run
command.

One still needs to define -d paranoid to get the checksumming, though.