Current state of the "chat-forth" world - a terrifying work-in-progress
And liable to change arbitrarily and irrevocably at any time. ;-)

I wanted to snapshot the current state of my thinking and
experimentation with this checkin.

I think the main lesson in all this is: support two stacks in the chat
(debug) firmware, otherwise it's tedious and ugly.

I'm not as sure that I need the forth-specific code that I've included -
it could just as well be in a debug-flavored kernel, or even in a few
bytes of ram that get uploaded at chat time.

But separating return addresses works really well - and not just for
forth.