Minor tweaks to chat and chat-peer
After all the head-scratching and conundrums that emerged from trying to
"tune" the chat code to work better with Forth code, I decided that the
main "innovation" in the chat-forth-* code was using two stacks, and
_switching_ them when going from the normal world of native code (like
the chat code itself) to the ITC Forth world, where SP is the data
stack, and RP is the return stack.

If you write code that keeps its state entirely in registers (as the
chat code does) you'll notice that the CPU's stack pointer (SP) register
really plays the role of Forth's RP (return stack pointer).

And if the host is trying to push stack items to the target, or pull
them from the target, this really only makes sense if you're using a
second stack - and a second stack pointer - to manage those data.

We then notice something strange: the CPU's SP is really RP; and we have
to use RP (an arbitrary register - on the MSP430 I've chosen r6) - as
SP.

But when we launch some Forth code, we have to switch these registers.
And when we return from Forth code to chat, we have to switch back.

But that is really the _only_ concession to the oddness of Forth that
has to be made in the chat code!

I was worried that because I couldn't change SP (because of the way I
specified the chat protocol!!) that I wouldn't be able to push data
items to the target's stack. But what I've now realised is that I
wouldn't be pushing them to that stack anway!! Instead, I'll want to
modify RP (really SP) in the stack frame to point to the items that I've
copied to the data stack area (below the return stack area).

When the Forth code starts up, it'll switch RP and SP, and SP will be
happily pointing to the stack items that the host has pushed. It is all
very simple and groovy.