New register model for msp chat
NOTE: Do not try to compile and flash this code yet!

Changes: top is no longer used by the chat code, and is no longer saved
on the stack frame!! We are assuming that it'll be a part of the data
stack that the host pushes (pointed to by RP).

Also, the registers have been re-ordered so that ones that the host is
likely to want to read and change - PC, SR, RP, IP - are together in a
contiguous chunk - and the registers that are used by chat, and saved so
as not to clobber them (even though they are almost all registers that
are considered by forth to be "scratch") are saved, but do not need to
be read or written by the host.

Overall, these are minor optimisations, rather than sweeping changes.

And until I try out my ideas about how the host is going to push and pop
the target stack (and top), and how the trampoline (for running forth
code) works, this code should just sit.

Currently the "newregs" code in chat-host and chat-peer has _NOT_ been
updated to match these changes!