Possibly figured out how to call forth from chat, and return again
We modify the chat code to only push _one_ return PC onto the stack.
When the host first connects, it grabs and saves this PC - which, if
jumped to, will re-enter chat, pushing a new frame (including the same
PC!).

By creating a small fake forth word in ram, we can call an arbitrary
piece of forth code and then re-enter chat.

It's not exactly pretty, but I think it will work.