Decided that it's best to keep both "outer" and "frame" PCs in chat
The reason: the outer one makes testing native code very easy: end the
routine with a return instruction, and you return to chat.

For Forth code we can't do this, because there might be some return
addresses on the stack, but if we simply ignore the outer return
address, and only use the "inner" (frame) return address to re-enter
chat - which we saved when we first connected to the target - then all
will be well.

Having both works. No need to choose.