Back to "falling back" to forth in __meta mode (when chatting)
Before I wanted to restrict "chatting" mode to .meta. and .target. words
only.

But there is a lot missing from that world, and it's a bit awkward at
the moment to get anything done.

So now the only difference between __building (the meta mode for
building the target image) and __chatting (the meta mode for interacting
with the target) is that __chatting tries to remotely execute .target.
words, before falling back on searching .forth.