Minor changes to msp430 interact and kernel
* codes added to meta

* nope added to kernel

* bug-to-chat is patched at chat connect time, not every time target
  forth code gets executed

* variables and buffers have their space fully allotted _before_ they
  are defined, so they will work if defined in ram (eg, during testing)

I'm trying to make it possible to execute a single word while in the
middle of a nested execution (ie, within bug within nested forth code).
I don't currently see how to do it. :-(

These changes were part of an attempt that didn't work, but they are
improvements that are useful regardless.