Solution A to problem of debugging while stopped after "bug"
After user code hits the word "bug", we will drop back into chat. If at
this point, another target forth is executed, unless we make special
provisions, this second execution will clobber the IP of the first, and
we will not be able to use "cont" to continue.

This commit solves this problem in the following way. It is based on the
observation that if user code is going to call bug, it can only do so in
a colon word, so the RP must change. Since SP in chat reflects RP in
target forth code, we save IP and SP before execution, then check
afterward if SP has changed; if it has _not_, we reset IP to the saved
value.

This seems to work, but will cause the IP (as shown in the register
printout) to behave oddly. If IP starts out as a garbage value, and if
the user only executes words that don't ever call bug, then IP will
always show as the same garbage value.

If the user ever calls a word that calls bug, and then uses "cont" to
run to completion, IP will end up at the "idle" value (right after "bug"
in trampoline) and will stay there after that. This could be confusing.

There is at least one other solution. I'm going to try that and check it
in as Solution B. Then we can choose.