Solution B to problem of debugging while stopped after "bug"
After user code hits the word "bug", we will drop back into chat. If at
this point, another target forth is executed, unless we make special
provisions, this second execution will clobber the IP of the first, and
we will not be able to use "cont" to continue.

This commit solves this problem in the following way. When we first chat
with the chip, we artificially set the IP to point to after the call to
"bug" in trampoline. I call this the "idle" IP - since there is a loop
there. Doing "cont" at the idle IP returns you to the idle IP.

Every time we execute a forth word using remote, we first save the IP.
If the execution completes - ie, ends up at the idle IP - we restore the
saved IP, in case it pointed into the body of another word that we were
already executing, but suspended by calling bug.

This is a bit of a hack, but it actually seems to work well. I think
it's a keeper. It works as well as Solution A, but improves on that by
having IP behave more consistently. Starting with the initial chat
connection, IP will mostly reflect the idle IP.