Save trims more intelligently
Before we were copying the trim from the chip as long it was _not_ ff.
But if we want to, for whatever reason, override or change the trims on
the chip, this becomes difficult!

This has been changed to do something arguably more sane: if the image
is 0ff, copy the corresponding byte from the chip, no matter what its
value. This works well.