Working code to drive Hitachi HD44780 LCD with MSP430!
I hadn't been careful enough about separately driving the "control"
lines (RS and R/nW) and the data lines, so things were getting garbled
on the bus.

I fixed the code, but it still didn't seem to work, but when I tried
chars in the range f0 to ff I got the right values... Then a1 showed as
"a" and a2 showed as "b"... I finally figured out that I had transposed
D7 and D6 when wiring to the chip!

After getting the wires right everything works fine!

I'm still not ecstatic about the code, though. It seems too long and
involved for something so conceptually simple.

And it's way too long. I think it consumes 340 bytes, which is waaay too
much. ;-)