HC08: region now uses current region, rather than expecting one
So, this started out as a small change, but then it blossomed into
something bigger.

I ended up needing/wanting to fix the craziness with the little bits of
firmware loaded into S08s and how they get mapped to addresses - my
latest hack of fslots didn't work out. It didn't help me with mapping
from slots to addresses - and there was still the distinction - now
erased - between USB addresses and serial addresses. The difference
existing because the USB-based BDM firware doesn't fit into 512 bytes,
so all the addresses were shifted on the USB devices.

I decided to make everything the same, even though it's a bit of a
waste. Actually, this whole thing is a bit of a waste: if I ran these
together into one chunk it would be much smaller, since the command loop
overhead keeps getting repeated, and many of the programs are only 128
bytes long or so, but consume a 512-byte flash sector.

Anyway, it's a mess, but it's slightly less of one now.