Make HC08 firmware map not depend on loaded chip
Since we might be using the HC08 to talk to an AVR or PIC, we need to
know the start addresses of the flash routines no matter what @flash
and #flash are.

And since we are not going to put this on a 908 - ever - just
assume/pretend that flash ends at "1_0000 - which it does for all the
S08's that we care about (ie, not the strange paged memory ones).