Simplified S08 flashing code
Now that I can write 4 KiB at a time over USB, I thought I'd improve and
simplify the S08 flashing code, which has been doing things in chunks of
64 or 128 bytes.

Now it both erases and programs a page at a time. Tested on the
bootloader and one other resident program.

I still need to test this with chat over serial.

Pretty pleased with it!