hc08: flash in chunks of 64 bytes - again
Even though we can do more via USB and via the serial chat interface,
flashing via BDM requires the use of the target stack to store the bytes
to be programmed. Since the RAM can be quite small - 256 or 512 bytes is
common - programming 512 bytes at a time obviously won't work.

Going back to chunks of 64 - which is what I had been doing for ages -
seems like a reasonable idea.