HC08 serial chat no longer sends status when it pushes its stack
This caused trouble, and was annoying. Sometimes - esp when returning
from another command loop back to chat - there would be a byte hanging
around in the serial input buffer on the host.

For the MSP430 chat code I decided that not having the Run command wait
for the target was a simpler and better design. Generally, for
interactive testing anyway, you're going to Run and then get the regs.
Well, when you fetch the regs you have to fetch the SP first, and that
command will stall until the target has returned to its command loop.

I made the same change to the HC08 chat, chat-host, and chat-peer code,
and added the option for chat-peer to emulate the "turds" behaviour of
the previous chat code.

Because chat does a lot of "resyncs", having the turds or not should be
fine. That is, old chats should work with new chat-hosts.