All HC08 serial command loops set baud rate and resync
On Linux tty devices - even USB ones - remember their settings from one
open to the next. This is not true on all systems! On BSD, for example,
USB ttys get reset to 9600 bps between openings, so each time we want to
chat with code on a serially-connected target we need to be careful
about setting the baud rate!