msp430: minor changes to flashing code and chat-peer
After erasing a page or flashing a chunk of words, the chat-host code
now resets the stack-resident FCMD to 0000. This way an inadvertent
programming of the flash is much less likely.

Also, the chat-peer code has been updated to re-zero the FCMD any time
code is "run" (by re-pushing the stack frame), and it also checks to see
if a valid FCMD is on the stack and returns a bogus flash status if not.

Testing this on real hardware would be nice too... ;-)