hc08: be more careful about flashing; clear fcmd each time
After each chunk or page erase clear the flash command so it doesn't
stay set as a valid command.

This way if a spurious FlashNext command comes down the wire, it's much
less likely to cause an accidental programming or erasure.

Also, the chat firmware clears the command byte each time it re-enters
the chat loop - which means it also happens at reset.