hc08: added chat-via, like in msp430; changed generic target cmds
These are both improvements, but add nothing in terms of features.

I liked how the use of chat-via (which basically sets a _single_ pointer
to a vector of routines for communicating with the target, rather than
setting a bunch of deferred words) worked out on the MSP430, so I
thought I'd make a similar change for the HC08/S08.

While I was at it, I simplified the word set used to communicate with
the underlying debug "transports".

Now, all reads and writes are chunked; no streaming (self-incrementing)
read/write at this level. The _underlying_ (eg, serial) chat code may
still do streaming reads and writes, but at the interface it gets
converted.

In one or two places (eg, save-trims in the S08 flashing code) got a bit
clunkier. But in general I think chunked read/write is faster and
simpler.

I converted the BDM code to use chat-via as well, but haven't tested it
yet. ;-)