Improved chat3 HC08 serial chat protocol!
I created chat3 a long time ago but never used it.

I've decided that I like the idea of a tail-recursive command-loopy
implementation for serial chat - it's how I do the USB chat and also all
the serial command loops for, eg, doing BDM or programming AVRs

Again, I thought, while I'm in there I might as well make it as simple
and pretty as possible.

And I took another idea from the MSP430 chat code - that of storing a
32-bit fragment of the muforth Git commit that defined the chat code
running on the chip. It's very easy to query and read back.

And, amazingly, the code is two bytes _shorter_ than the old chat code.
;-)