hc08: chat3 is now the standard serial chat protocol!
What this means for you:

If you are running an S08 board over serial and don't want to change
anything, you have to remember to add

  -d chat2-host

to your command line (before any -f target/HC08/<blah>) in order to
continue using chat2.

However, it's easy peasy to switch over. Here are the steps:

  ./muforth -d in-ram -d chat2-host -f target/HC08/load-chat.mu4

This will load the chat3 code into RAM, but use the chat2 protocol on
the host-side to talk to the _existing_ chat2 code on the chip.

Then do:

  chat
  go

This will start the new chat3 code in RAM. Your PC should change to
something like 0182 from fe71 (or so).

Now quit muforth and restart with the following command:

  ./muforth -f target/HC08/load-chat.mu4

This will load the chat3 flash-resident code, using chat3-host as well
(to talk to the RAM-resident chat3 code).

Now:

  flash-image
  verify

And then power cycle the chip to get back into chat. You'll see, among
other things, a nice commit telling you what version of muforth the chat
code is based on (not that it's ever going to change! ;-)

Now that this is committed, I'm going to change all my chips over so
they'll show the right and proper commit.