Converted to 3-clause BSD license from 2-clause
And aren't *you* lucky!! I almost changed muforth to use the GPL but
after thinking about it (for three years!) I decided that I still liked
the BSD license, even though some evil company might take my software
and run with it...

The fact is that it's much easier, using a BSD license, to take code
with you, both *into* and *out of* a company, if it's BSD-licensed. In
other words, because of the commercially-friendly nature of the license,
a company that you want to work for may choose to use *your* code and
hire you to work on it, and if/when you leave, you can take the code
with you (assuming that you didn't sign an agreement that your work is
the copyright of your employer).

The GPL offers much less freedom in this regard.