Still more changes to USB enumeration under Linux...
I decided to be even more careful about errors. I was looking at the
FreeBSD and NetBSD USB code (for muFORTH) and decided that there *is* an
error condition that we want to report: when the vid and pid of a device
match that requested, but re-opening the device read/write fails. We'd
like to inform the user of this fact.

Sooo, the code is now very careful to return -1 (error) only in that one
case.