Number parsing is now independent of 'complain'
Before, if a token couldn't be parsed as a number, complain was
automatically called. It prints the token, the message "isn't defined"
and then *aborts*, which unwinds the R stack. This is sometimes *not*
what you want!

The behaviour of the word  number  stays the same. It calls complain if
the token can't be parsed as a number.

The new wrinkle is  number?  It tries to convert a token as a number,
and leaves a flag - true for success, false for failure. This way number
conversion can be more easily embedded into other code, without having
to worry about unscheduled aborts!