ARM: Scaffolding for tini2131 - load files, etc
Getting ready to undertake the LPC bootloader code. I exhumed some code
I wrote long enough ago that the comment says "Philips LPC2100" rather
than NXP LPC2100! Any historians out there?

This code was a strange beginning of an approach - that I rather like -
to the quirky LPC bootloader, which wants to have a conversation with
the host. So I set up a vocab chain with some words in it pertaining to
the LPC protocol. I'm not sure how much I can keep, and right now it
does nothing, but it's here, as a start.