MSP430: reset command waits between Run and GetRegs commands
When doing a reset command, esp when running the flash-based chip
startup code, not only do some of the initializations take time, but
setting up the UART involve *resetting* it, which loses any pending
bytes in the buffer. Our previous code had sent a "get regs" command
immediately after the Run command, but the UART reset had thrown it out.

Delaying for a bit (100ms) gives the chip time to get going before we
request the registers.

This makes switching between flash-based and ram-based chat (used to
re-flash the flash-based chat code) more seamless. (Before you would get
a "reset timed out" but if you said "chat" again everything would be
fine, so it was confusing and did not inspired confidence in the
proceedings!)

Speaking of re-flashing the chat code... We need to do that to bring
back paranoid mode. Here's how:

  ./muforth -d in-ram -f target/MSP430/load-chat.mu4
  chat
  reset
  ^C

You should have seen the PC switch from fefe (or thereabouts) to 0330.

  ./muforth -f target/MSP430/load-chat.mu4
  chat
  flash-image
  verify
  reset

And you should be running the new code. -d paranoid will now work again.
;-)