MSP430: Further improvements to paranoid chat
The reset command would break in paranoid mode when resetting back to
the flashed chat firmware. The reason: UART reset and the time to set up
the chip clock, etc. Same thing. But this time, instead of trying to
read the registers immediately after the Run command, we were trying to
get the chat firmware's accumulated checksum.

Now we wait until we are fetching the registers. This way the new 100ms
delay in the interact code can kick in and help us out.