MSP430: chat (and remote execution!) has been totally revamped!
Big changes.

First, the basic chat firmware has been rewritten with two aims in mind:

* make the chat firmware totally generic and not tied to a particular
  model (eg, of ITC Forth) for remote execution of code;

* change the model of register use and how registers are saved and
  restored, to be simple and minimal.

The chat code was rewritten to only use four registers (instead of
five), and those registers were all taken from the "top" of the
register set (12 to 15), which, by convention, we are treating as
"caller-saved" (scratch) registers. The only registers pushed onto the
chat stack frame and exchanged with the connected host machine are PC,
SP, SR (status register), and a register we are calling CP, the context
pointer (r4). CP can be used by code on both the target and host to
exchange more context (such as Forth VM state, Forth data stack values,
etc) than the generic chat code.

The Forth execution part of the story, which was a bit of a tangle, has
actually been simplified by this change. Using the "pause" and "run"
words of a multitasker as a model for how Forth VM state is saved and
restored, it became clear that simply pushing VM registers (IP mostly)
onto the data stack, and then copying the stack to the host (and
reversing the process when beginning execution of a word) was the
simplest way to set up and monitor remote execution of Forth words.

The results actually exceeded my expectations, in both simplicity and
clarity.

Another change is in how *starting* and *continuing* execution are
distinguished. It was, and is, possible to execute a word that somewhere
during its execution calls the word "bug", which re-enters the
chat/debug conversation with the host. But since the execution of the
original word hasn't completed, we need a way both to indicate this, and
to allow the user to "continue" execution.

When a Forth word is initially executed, the target IP is set to a
particular value (the beginning of a word defined in the kernel, called
"trampoline"). trampoline is very simple: it executes the word whose cfa
is on the stack, and then calls "bug", and if the user continues at this
point, "bug" is immediately called again. So the obvious test for
"completion" of execution is whether the IP currently points just past
the call to "bug" in trampoline. If this is true, we print a normal
register set. But if this *isn't* true, we are probably in the middle of
executing a word (or words) with embedded calls to "bug". In this case
we annotate the IP address with a following "*", suggesting that the
user should run "cont" to continue execution.

The only distinction between starting and continuting execution of Forth
code is whether or not IP is set (execute a new word), or simply left
alone (cont).

Since this change involves a new version of the chat firmware - and a
new chat protocol to boot - the chat firmware needs to be reflashed, but
it's a tiny bit more involved than usual because of the protocol change.
In the following command, we do "-d chat-v1" so that we can talk to the
*existing* flash-based chat code; "-d in-ram" will load the *new* (v2)
chat into ram, and "go" will jump to it. So, to update, do this (-d
paranoid is optional but should work - however, I didn't actually test
using it during reflashing):

  ./muforth -d in-ram -d chat-v1 -f target/MSP430/load-chat.mu4
  chat
  go
  ^C

You are now running the new v2 chat code in RAM.

  ./muforth -f target/MSP430/load-chat.mu4
  chat
  flash-image
  verify
  reset

And you should be running the new code. PC will change to something like
ff30, and the register printout should look much different (it's much
shorter - just status, CP, SP, and PC).

To try out the new kernel goodness, do this (after flashing the new
chat):

  ./muforth -f target/MSP430/load-kernel.mu4

Try executing a few words, like "0", "-1 negate", "-4 5 +".

Then, for some real fun, try this:

  ram
  : dd   0 do  i bug drop  loop ;
  5 dd

This will start execution, but because of the embedded bug, it will not
finish, but will re-enter chat. The current index will be on the stack
(because of "i"), and the internal count will be in the Cnt register.
Type "cont" to continue. Keep doing this until the "*" annotation on IP
disappears.

This is the new way. ;-)