Trying to debug the CMSIS-DAP code...
Again, this is months old. But I remember having weird issues with the
CMSIS-DAP code not being very reliable, so I had started to make some
changes to try to instrument it better to see what was happening.

This is a work in progress. I'm checking it in because I really want to
update the copyright to 2016 - now that it's June!