WIP - porting target/V25/ to muforth!
I got stuck in the metacompiler because I need a feature that muforth
used to have - a hidden bit in the name field - but now lacks. I've been
thinking of adding it back (and the ability to link vocab chains) and
thought I'd check in what I've got so far for safekeeping, while I look
into bringing back some past features.