WIP: MSP430 DCO calibration
Right now it's just a few routines I can use to understand how the
timers on the MSP430 work. I was able to capture a 4kHz square wave,
generated by another chip, and use that to tune the DCO to be closely
matched. I was doing this by hand; I want to write code that will do it
automgically.

There is example code from TI that does just this. I'm in the middle of
modifying and adapting it!