MSP430: minor changes to Forth kernel
I decided to opt for the slightly simpler and slightly slower version of
next - one that auto-increments the w register in next. While this takes
one cycle longer (5 rather than 4), it simplifies and shortens (by one
instruction and one cycle) all of the Forth runtimes: dodoes, docolon,
doconstant, etc.

I neglected at first to fix execute - the last instruction of which is
the same as second part of next, and likewise needs to auto-increment w
- so my tests showed odd behavior: constants were pushing their code
  address (c010) rather than their value.

Once execute was fixed, everything worked fine.

I also added  ram  to the end of load-kernel, so when doing experiments
one is compiling into RAM rather than flash...