MSP430: new register conventions, two kinds of looping words
I've added versions of for/next and do/loop that keep the loop index in
the stack rather than in a register. This is a tiny bit slower, but the
code is also slightly shorter, and in a multi-tasking system there is
one less Forth VM register to push and pop at every task switch.

The interaction experience is the same regardless of the implementation.
The current loop index - which counts down for for loops, and usually
counts up for do loops - unless you are using +loop with a negative
offset - are shown for each interation in the Loop register.

To continue to build the loop-in-register version, change nothing.

To build the loop-on-stack version, do something like this:

  ./muforth -d loop-on-stack -f target/MSP430/load-kernel.mu4
  chat
  flash-image
  verify

and from then on, wheneever you load the kernel code - directly or
indirectly - you have to specify the -d loop-on-stack to get the correct
behaviour.