Colorforth display now uses improved console colors
With the default colors, display of numbers and macros was difficult to
read. Numbers - which are "bright" when hexadecimal - looked bad when
bright, so I changed this to use a muforth-style " prefix. And macros,
which, when compiled rather than called are colored cyan, were hard to
distinguish from green.

With the improved console colors, both these problems are solved, and
display with the new colors looks pretty great - esp using a big 16x32
font like Terminus!