Brought ARM/v5 code up-to-date with 64bit muforth and untoken
I wanted to see how the "writing out the target image as a C array" code
worked but had to massage a few things to get it working again with a
64bit version of muforth. (A few places assumed that both the host and
target cell sizes were 32 bits.)

Also, I've since removed the multiple create words that have both
parenthetical versions (create) (constant) and non-paren versions -
create, constant - the latter consuming a token from the input stream
and the former expecting an (a u) on the stack.

Now I just use  untoken  to back up over the last token read. This is much
simpler!!