msp430: fixed pc-relative bugs in asm and disasm
I also clarified (I hope) some code and comments.

Among other things, the "optimization" to convert

  0 <reg> +)   to   <reg> )

caused all kinds of havoc, including totally breaking the (so far
unused) pc-relative addressing mode code.

I had to fix the offsetting as well, both in the asm and disasm. I was
offsetting from the address of the offset rather than the address of the
following instruction.