msp430: reverted broken "fix" of pc-relative asm and disasm changes
This partially reverts commit 4123b0c44b33e4620a5829926153e5a4893a16bd

After a recent perusal of the "family reference manual" for the
MSP430G2553 chips I'm using to write and test this code, I realized that
the previous versions of the assembler and disassembler were correct in
how they treated pc-relative offsets, and that commit broke them.

The issue in cases like this is always this: Where does the PC when
doing the calculation of adding in the offset? For relative jumps, on
this architecture as on many others, the PC points *at the start of the
next instruction*. I assumed, in the above commit, that this was also
true for PC-rel displacements. But it is not so.

For one thing, there can be more than one. In fact, there can be two!
Both the source and dest can be PC-rel. So the chip designers, either to
simplify the life of the compiler and assembler writer or their own,
decided that for PC-relative data offsets, that the PC points *at* the
offset, not somewhere past it.

Next I want to test the behaviour of "add pc, #offset" - that doesn't
seem to be documented anywhere.