msp430: Added simple test code for PC-relative addressing
I wanted to verify the behaviour, of both documented (PC-relative
addressing of operands) and undocumented (add #offset, PC) instructions.

Conclusions:

* add #offset, PC uses the address of the next instruction as the PC, no
matter how long the instruction is. I tried this with constant-generator
immediates and true immediates and got the same results.

* PC-relative addressing behaves as documented in the instruction set
chapter of the family manual: offsets are added to their own address,
not the address of the following instruction.

The assembler and disassembler do the right thing in the second case. In
the first case you are on your own to calculate the correct immediate
value to add to the PC.