msp430: for/next loops are now always "smart":  0 for  loops 0 times
It used to be that get this behaviour - passing a 0 count to a for/next
and having it loop times - required code like this:

  ?for  blah blah blah  next  then

A regular for loop like

  for  blah blah blah  next

would execute 2^wordsize times if given a 0 count. This is never what
you want, so having for loops always be smart seems ... smart. So that
is what we have now.

Under the hood, here is what happens:

* for compiles the following code:
  (?0branch) <address slot> (for)

  where
     (?0branch) is a branch runtime that branches and leaves the value on
                the stack if it's nonzero, but if zero, it pops it and skips
                the branch

     (for)      is the for loop runtime, which sets up the loop counter -
                either in a register or in a slot on the return stack
                (depending on how the kernel code was compiled)

* next compiles the following code:
  (next) <backward-branch-address>
  and then also resolves the forward branch that for compiled.

Anyway, regardless of the fiddly details, this is a much better world.
So much better that the next thing I'm going to do is probably retrofit
this into the host muforth!

BTW: though the msp430 kernel file changed, the changes were all
cosmetic. No code changed, so no re-flashing is necessary.

Yet. ;-)